services/windower/requirements.txt
flask>=3.0
gunicorn>=21.0
pandas>=2.0
numpy>=1.24
pyarrow>=12.0
google-cloud-storage>=2.10
google-cloud-pubsub>=2.18
